Gen AI 2.O/MCP/azure-blob/requirements.txt
mcp[cli]
azure-storage-blob
